moved components folder inside app directory
