+++ app/components/Navbar/desktop/index.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import ThemeSwitch from '@/components/Theme';
+import ThemeSwitch from '@/app/components/Theme';
 
 import NavMenuItems from '../NavMenuItems';
 import DownloadIcon from '../mobile/DownloadIcon';
