+++ app/components/Navbar/mobile/index.tsx
@@ -2,7 +2,7 @@
 import { useState } from 'react';
 import {AnimatePresence} from 'framer-motion';
 
-import ThemeSwitch from '@/components/Theme';
+import ThemeSwitch from '@/app/components/Theme';
 
 import HamburgerIcon from './HamburgerIcon';
 import MobileMenu from './MobileMenu';
